+++ lib/grape/validations/validators/values.rb
@@ -7,7 +7,7 @@ module Grape
       end
 
       def validate_param!(attr_name, params)
-        return unless (params[attr_name] || required_for_root_scope?) 
+        return unless params[attr_name] || required_for_root_scope?
 
         values = @values.is_a?(Proc) ? @values.call : @values
         param_array = params[attr_name].nil? ? [nil] : Array.wrap(params[attr_name])